+++ src/library/DIPS.Mobile.UI/Components/Pickers/DatePicker/Android/DatePickerHandler.cs
@@ -1,8 +1,9 @@
 using System.Globalization;
 using Android.Content.Res;
+using DIPS.Mobile.UI.Components.Chips;
+using DIPS.Mobile.UI.Components.Chips.Android;
 using DIPS.Mobile.UI.Components.Pickers.DatePicker.Android;
 using DIPS.Mobile.UI.Converters.ValueConverters;
-using DIPS.Mobile.UI.Platforms.Android;
 using Microsoft.Maui.Handlers;
 using Chip = Google.Android.Material.Chip.Chip;