+++ src/library/DIPS.Mobile.UI/Components/Pickers/TimePicker/Android/TimePickerHandler.cs
@@ -1,7 +1,8 @@
 using System.Globalization;
+using DIPS.Mobile.UI.Components.Chips;
+using DIPS.Mobile.UI.Components.Chips.Android;
 using DIPS.Mobile.UI.Components.Pickers.TimePicker.Android;
 using DIPS.Mobile.UI.Converters.ValueConverters;
-using DIPS.Mobile.UI.Platforms.Android;
 using Microsoft.Maui.Handlers;
 using Chip = Google.Android.Material.Chip.Chip;