+++ src/library/DIPS.Mobile.UI/Extensions/Markup/StringFormatExtension.cs
@@ -1,0 +1,19 @@
+namespace DIPS.Mobile.UI.Extensions.Markup;
+
+[ContentProperty(nameof(Format))]
+public class StringFormatExtension : IMarkupExtension
+{
+    /// <summary>
+    /// A composite format string.
+    /// </summary>
+    public string Format { get; set; }
+    /// <summary>
+    /// The object to format.
+    /// </summary>
+    public string Argument { get; set; }
+    
+    public object ProvideValue(IServiceProvider serviceProvider)
+    {
+        return string.Format(Format, Argument);
+    }
+}